MVP Achitecture in LoanChargesDialog
