+++ mifosng-android/src/main/java/com/mifos/mifosxdroid/injection/component/ActivityComponent.java
@@ -6,6 +6,7 @@ import com.mifos.mifosxdroid.dialogfragments.datatablerowdialog.DataTableRowDial
 import com.mifos.mifosxdroid.dialogfragments.documentdialog.DocumentDialogFragment;
 import com.mifos.mifosxdroid.dialogfragments.loanaccountapproval.LoanAccountApproval;
 import com.mifos.mifosxdroid.dialogfragments.loanaccountdisbursement.LoanAccountDisbursement;
+import com.mifos.mifosxdroid.dialogfragments.loanchargedialog.LoanChargeDialogFragment;
 import com.mifos.mifosxdroid.injection.PerActivity;
 import com.mifos.mifosxdroid.injection.module.ActivityModule;
 import com.mifos.mifosxdroid.login.LoginActivity;
@@ -123,4 +124,6 @@ public interface ActivityComponent {
     void inject(LoanAccountApproval loanAccountApproval);
 
     void inject(LoanAccountDisbursement loanAccountDisbursement);
+
+    void inject(LoanChargeDialogFragment loanChargeDialogFragment);
 }
